Revert "Cleaning up Update User Password, and Validate Two Factor (#573)"
This reverts commit a3aefbeb1f76b72973ab96e1ff933a8733c33775.
