+++ frontend/schema.faker.graphql
@@ -33,7 +33,7 @@ that gives them access to restricted content.
 """
 type Authenticate {
   """
-  user info who just signed in, and their JWT
+  User info who just signed in, and their JWT
   """
   authResult: AuthResult
 }
@@ -48,7 +48,7 @@ type AuthResult {
   authToken: String
 
   """
-  user that has just been created or signed in
+  User that has just been created or signed in
   """
   user: User
 }
@@ -852,7 +852,7 @@ type Mutation {
     password: String!
 
     """
-    user email that they signed up with.
+    User email that they signed up with.
     """
     userName: EmailAddress!
   ): Authenticate
@@ -1793,12 +1793,12 @@ type UserAffClass implements Node {
   id: ID!
 
   """
-  user's ID
+  User's ID
   """
   userId: Int
 
   """
-  user's level of access to a given organization
+  User's level of access to a given organization
   """
   permission: RoleEnums
 }
