+++ frontend/src/__tests__/UserList.test.js
@@ -32,7 +32,7 @@ describe('<UserList />', () => {
                     userName: 'testuser@testemail.gc.ca',
                     admin: true,
                     tfa: false,
-                    displayName: 'Test user Esq.',
+                    displayName: 'Test User Esq.',
                   },
                 },
               ],
@@ -92,7 +92,7 @@ describe('<UserList />', () => {
                     userName: 'testuser@testemail.gc.ca',
                     admin: true,
                     tfa: false,
-                    displayName: 'Test user Esq.',
+                    displayName: 'Test User Esq.',
                   },
                 },
               ],
