+++ platform/overlays/gke/cluster-admin-rolebinding.yaml
@@ -9,5 +9,5 @@ roleRef:
   name: cluster-admin
 subjects:
 - apiGroup: rbac.authorization.k8s.io
-  kind: user
+  kind: User
   name: mike.williamson@gcdigital.canada.ca
